Removed unnecessary type casting.

+++ src/Walkers/WalkerDrift.java
@@ -69,7 +69,7 @@ public class WalkerDrift extends Walker implements UnitInterface
                     g = grid.getGrass(this.getX()+(x-1), this.getY()+(y-1));
                     double angleg = angle(g.getX(), g.getY());
                     //System.out.println("x:"+g.getX()+"y:"+g.getY()+"="+angleg + "<" + Math.PI / 2);
-                    if (angleg < (double)(Math.PI / 3))
+                    if (angleg < (Math.PI / 3))
                     {
                         //System.out.println("TRUE");
                         posTar.add(g);
